handwriting_ocr_eng/requirements.txt
bdrk==0.2.2
numpy==1.19.1
gluonnlp==0.9.1
leven==1.0.4
mxboard==0.1.0
mxnet==1.6.0
opencv-python==4.2.0.32
pillow==7.0.0
sacremoses==0.0.38
scikit-image==0.16.2
streamlit==0.56.0
